refactoring around branch_utils
