+++ README.md
@@ -134,6 +134,8 @@ You can get additional help information for a particular command by specifying *
        -b, --branch <branch>            Specify the name of remote git branch. If not provided, we will try to
                                         automatically detect the current working branch if the command is used in a
                                         git enabled folder.
+                                        In cases branch is not specified AND local state has uncommitted changes or 
+                                        untracked files a temporary branch will be created and used.
 
        -c, --commit <commitId>          Specify commit ID. It's required to run sandbox from a blueprint from an
                                         historic commit. Must be used together with the branch option.
